refactor: Type casting 책임 배분
