+++ src/main/java/blackjack/domain/player/Participant.java
@@ -19,6 +19,13 @@ public class Participant extends Player {
         }
     }
 
+    public static Participant changeToParticipant(Player player) {
+        if (player instanceof Participant) {
+            return (Participant) player;
+        }
+        throw new ClassCastException("[ERROR] Player가 참여자가 아닙니다.");
+    }
+
     public void win() {
         this.winState = Result.WIN;
     }
